@@ -1,8 +1,6 @@
 #pragma once
 #include "../Special.h"
 
-#include <array>
-
 namespace xxx::refl
 {
     template <typename T, size_t N>
@@ -14,21 +12,18 @@ namespace xxx::refl
     class StdArray : public Special
     {
     public:
-        virtual SpecialType getSpecialType() const
-        {
-            return SpecialType::Std_Array;
-        }
-        virtual void setElementValue(void* instance, size_t index, void* data) const = 0;
         virtual Type* getElementType() const = 0;
         virtual size_t getElementCount() const = 0;
-        virtual void* getElementPtrByIndex(void* instance, size_t index) const = 0;
+        virtual void setElement(void* instance, size_t index, const void* data) const = 0;
+        virtual void getElement(const void* instance, size_t index, void* data) const = 0;
+        virtual void* getElementPtr(void* instance, size_t index) const = 0;
+        virtual const void* getElementPtr(const void* instance, size_t index) const = 0;
 
     protected:
-        StdArray(std::string_view name, size_t size) : Special(name, size) {}
+        StdArray(std::string_view name, size_t size) : Special(name, size, Case::StdArray) {}
     };
 
-    class Reflection;
-    template <typename T, typename = std::enable_if_t<is_std_array_v<T>>>
+    template <typename T> requires is_std_array_v<T>
     class TStdArray : public StdArray
     {
         friend class Reflection;
@@ -60,11 +55,6 @@ namespace xxx::refl
             return false;
         }
 
-        virtual void setElementValue(void* instance, size_t index, void* data) const override
-        {
-            static_cast<std::array<Element, Count>*>(instance)->at(index) = *(Element*)(data);
-        }
-
         virtual Type* getElementType() const override
         {
             return Reflection::getType<Element>();
@@ -75,19 +65,33 @@ namespace xxx::refl
             return Count;
         }
 
-        virtual void* getElementPtrByIndex(void* instance, size_t index) const override
+        virtual void setElement(void* instance, size_t index, const void* data) const override
         {
-            return &(static_cast<std::array<Element, Count>*>(instance)->at(index));
+            static_cast<std::array<Element, Count>*>(instance)->at(index) = *(const Element*)(data);
         }
 
-    protected:
-        std::string_view genName() const
+        virtual void getElement(const void* instance, size_t index, void* data) const override
+        {
+            *(Element*)(data) = static_cast<const std::array<Element, Count>*>(instance)->at(index);
+        }
+
+        virtual void* getElementPtr(void* instance, size_t index) const override
+        {
+            return &static_cast<std::array<Element, Count>*>(instance)->at(index);
+        }
+
+        virtual const void* getElementPtr(const void* instance, size_t index) const override
+        {
+            return &static_cast<const std::array<Element, Count>*>(instance)->at(index);
+        }
+
+    private:
+        static std::string_view getName()
         {
             static std::string name = "std::array<" + std::string(Reflection::getType<Element>()->getName()) + ", " + std::to_string(Count) + ">";
             return name;
         }
 
-        TStdArray() : StdArray(genName(), sizeof(std::array<Element, Count>)) {}
-        virtual ~TStdArray() = default;
+        TStdArray() : StdArray(getName(), sizeof(std::array<Element, Count>)) {}
     };
 }
