@@ -1,8 +1,6 @@
 #pragma once
 #include "../Special.h"
 
-#include <unordered_set>
-
 namespace xxx::refl
 {
     template <typename T>
@@ -13,24 +11,18 @@ namespace xxx::refl
     class StdUnorderedSet : public Special
     {
     public:
-        virtual SpecialType getSpecialType() const
-        {
-            return SpecialType::Std_Set;
-        }
-
         virtual Type* getElementType() const = 0;
         virtual size_t getElementCount(const void* instance) const = 0;
         virtual std::vector<const void*> getElementPtrs(const void* instance) const = 0;
-
-        virtual void insertElement(void* instance, void* element) const = 0;
-        virtual void removeElement(void* instance, void* element) const = 0;
+        virtual void insertElement(void* instance, const void* element) const = 0;
+        virtual void removeElement(void* instance, const void* element) const = 0;
 
     protected:
-        StdUnorderedSet(std::string_view name, size_t size) : Special(name, size) {}
+        StdUnorderedSet(std::string_view name, size_t size) : Special(name, size, Case::StdUnorderedSet) {}
     };
 
     class Reflection;
-    template <typename T, typename = std::enable_if_t<is_instance_of_v<T, std::unordered_set>>>
+    template <typename T> requires is_instance_of_v<T, std::unordered_set>
     class TStdUnorderedSet : public StdUnorderedSet
     {
         friend class Reflection;
@@ -58,7 +50,8 @@ namespace xxx::refl
 
         virtual bool compare(const void* instance1, const void* instance2) const override
         {
-            if (static_cast<const std::unordered_set<Element>*>(instance1)->size() == 0 && static_cast<const std::unordered_set<Element>*>(instance2)->size() == 0)
+            if (static_cast<const std::unordered_set<Element>*>(instance1)->size() == 0 &&
+                static_cast<const std::unordered_set<Element>*>(instance2)->size() == 0)
                 return true;
             return false;
         }
@@ -75,25 +68,24 @@ namespace xxx::refl
 
         virtual std::vector<const void*> getElementPtrs(const void* instance) const override
         {
-            std::vector<const void*> result;
             const std::unordered_set<Element>* unorderedSet = static_cast<const std::unordered_set<Element>*>(instance);
-            for (auto it = unorderedSet->begin(); it != unorderedSet->end(); ++it)
-            {
-                result.emplace_back(&(*it));
-            }
+            std::vector<const void*> result(unorderedSet->size());
+            size_t i = 0;
+            for (const Element& element : *unorderedSet)
+                result[i++] = &element;
             return result;
         }
 
-        virtual void insertElement(void* instance, void* element) const override
+        virtual void insertElement(void* instance, const void* element) const override
         {
             std::unordered_set<Element>* unorderedSet = static_cast<std::unordered_set<Element>*>(instance);
-            unorderedSet->insert(*(Element*)(element));
+            unorderedSet->insert(*(const Element*)(element));
         }
 
-        virtual void removeElement(void* instance, void* element) const override
+        virtual void removeElement(void* instance, const void* element) const override
         {
             std::unordered_set<Element>* unorderedSet = static_cast<std::unordered_set<Element>*>(instance);
-            auto findResult = unorderedSet->find(*(Element*)(element));
+            auto findResult = unorderedSet->find(*(const Element*)(element));
             if (findResult != unorderedSet->end())
                 unorderedSet->erase(findResult);
         }
