@@ -1,5 +1,6 @@
 #pragma once
 #include "Serializer.h"
+
 #include <Engine/Utility/Guid.h>
 #include <stack>
 
@@ -7,184 +8,141 @@ namespace xxx
 {
     class Asset;
 
-    enum AssetSerialFlag : uint32_t
-    {
-
-    };
-
-    class ObjectBuffer
-    {
-    public:
-        ObjectBuffer(uint32_t bufferSize = 0, uint32_t pointer = 0) :
-            mBuffer(bufferSize), mPointer(pointer) {}
-
-        inline uint32_t tell() const
-        {
-            return mPointer;
-        }
-
-        inline bool seek(uint32_t pos)
-        {
-            if (pos <= mBuffer.size())
-            {
-                mPointer = pos;
-                return true;
-            }
-            return false;
-        }
-
-        inline uint32_t getSize() const
-        {
-            return mBuffer.size();
-        }
-
-        inline const void* getData() const
-        {
-            return mBuffer.data();
-        }
-
-        inline void writeData(const void* data, uint32_t size)
-        {
-            uint32_t newPointer = mPointer + size;
-            if (newPointer > mBuffer.size())
-                mBuffer.resize(newPointer);
-            std::memcpy(mBuffer.data() + mPointer, data, size);
-            mPointer = newPointer;
-        }
-
-        inline void readData(void* data, uint32_t size)
-        {
-            std::memcpy(data, mBuffer.data() + mPointer, size);
-            mPointer += size;
-        }
-
-    protected:
-        std::vector<uint8_t> mBuffer;
-        uint32_t mPointer;
-    };
-
-    struct ImportItem
-    {
-        Guid objectGuid;
-        uint32_t pathStrIndex;
-    };
-
-    struct ExportItem
-    {
-        Guid objectGuid;
-        uint32_t classNameStrIndex;
-    };
-
-    struct AssetHeader;
     class AssetSerializer : public Serializer
     {
     public:
         AssetSerializer(Asset* asset) : mAsset(asset) {}
         virtual ~AssetSerializer() = default;
 
-        // serialize from root object
-        virtual void serialize(Object** object) override
-        {
-            serializeClass(object);
-        }
+        virtual void serializeFundamental(refl::Fundamental* fundamental, void* data, size_t count = 1) override;
 
-        void fillAssetHeader(AssetHeader& header);
+        virtual void serializeStructure(refl::Structure* structure, void* data, size_t count = 1) override;
+
+        inline Asset* getAsset() const
+        {
+            return mAsset;
+        }
 
         uint32_t addString(const std::string& str);
 
-        uint32_t addImportItem(const ImportItem& importItem);
-
-        uint32_t addExportItem(const ExportItem& exportItem);
-
-        uint32_t createNewObjectBuffer(uint32_t bufferSize = 0);
-
-        inline const std::vector<std::string>& getStringTable() const
+        inline const std::string& getString(uint32_t strIndex) const
         {
-            return mStringTable;
+            return mStringTable[strIndex];
         }
 
-        inline const std::vector<ImportItem>& getImportTable() const
+        inline uint32_t getStringTableSize() const
         {
-            return mImportTable;
+            return mStringTable.size();
         }
 
-        inline const std::vector<ExportItem>& getExportTable() const
+        uint32_t addImportItem(uint32_t importItem);
+
+        inline uint32_t getImportItem(uint32_t importIndex) const
         {
-            return mExportTable;
+            return mImportTable.at(importIndex);
         }
 
-        inline const std::vector<ObjectBuffer>& getObjectBufferTable() const
+        inline uint32_t getImportTableSize() const
         {
-            return mObjectBufferTable;
+            return mImportTable.size();
+        }
+
+        class ObjectBuffer : public osg::Referenced
+        {
+        public:
+            ObjectBuffer(size_t bufferSize = 0) :
+                mBuffer(bufferSize), mPointer(0) {}
+
+            inline size_t tell() const
+            {
+                return mPointer;
+            }
+
+            inline bool seek(size_t pos)
+            {
+                if (pos <= mBuffer.size())
+                {
+                    mPointer = pos;
+                    return true;
+                }
+                return false;
+            }
+
+            inline void write(const void* data, size_t size)
+            {
+                uint32_t newPointer = mPointer + size;
+                if (newPointer > mBuffer.size())
+                    mBuffer.resize(newPointer);
+                std::memcpy(mBuffer.data() + mPointer, data, size);
+                mPointer = newPointer;
+            }
+
+            inline void read(void* data, size_t size)
+            {
+                std::memcpy(data, mBuffer.data() + mPointer, size);
+                mPointer += size;
+            }
+
+            uint8_t* getData()
+            {
+                return mBuffer.data();
+            }
+
+            size_t getSize() const
+            {
+                return mBuffer.size();
+            }
+
+        private:
+            std::vector<uint8_t> mBuffer;
+            size_t mPointer;
+        };
+
+        inline ObjectBuffer* getCurrentObjectBuffer()
+        {
+            if (mObjectBufferIndexStack.empty())
+                return nullptr;
+            uint32_t objectBufferIndex = mObjectBufferIndexStack.top();
+            return mObjectBuffers.at(objectBufferIndex);
+        }
+
+        inline void pushObjectBuffer(uint32_t objectBufferIndex)
+        {
+            mObjectBufferIndexStack.push(objectBufferIndex);
+        }
+
+        inline void popObjectBuffer()
+        {
+            if (!mObjectBufferIndexStack.empty())
+                mObjectBufferIndexStack.pop();
+        }
+
+        inline uint32_t createObjectBuffer(size_t bufferSize = 0)
+        {
+            mObjectBuffers.emplace_back(new ObjectBuffer(bufferSize));
+            return mObjectBuffers.size() - 1;
+        }
+
+        inline uint32_t getObjectBufferCount() const
+        {
+            return mObjectBuffers.size();
         }
 
     protected:
-        // using in AssetSaver
-        int32_t getIndexOfObject(Object* object);
+        virtual void serializeBinary(void* data, size_t size) = 0;
 
-        // using in AssetLoader
-        Object* getObjectByIndex(int32_t index);
-
-        inline void pushObjectBufferIndex(uint32_t index)
-        {
-            mObjectBufferIndexStack.push(index);
-        }
-
-        inline void popObjectBufferIndex()
-        {
-            mObjectBufferIndexStack.pop();
-        }
-
-        inline uint32_t tell() const
-        {
-            return getCurrentObjectBuffer().tell();
-        }
-
-        inline bool seek(uint32_t pos)
-        {
-            return getCurrentObjectBuffer().seek(pos);
-        }
-
-        inline ObjectBuffer& getCurrentObjectBuffer()
-        {
-            return mObjectBufferTable.at(mObjectBufferIndexStack.top());
-        }
-
-        inline const ObjectBuffer& getCurrentObjectBuffer() const
-        {
-            return mObjectBufferTable.at(mObjectBufferIndexStack.top());
-        }
-
-        inline bool currentObjectBufferIsValid() const
-        {
-            return !mObjectBufferIndexStack.empty();
-        }
-
-        template <typename T, typename = std::enable_if_t<std::is_arithmetic_v<T>>>
-        inline void serializeArithmetic(T* data, size_t count = 1)
+        template <typename T> requires std::is_arithmetic_v<T>
+        void serializeArithmetic(T* data, size_t count = 1)
         {
             serializeBinary(data, sizeof(T) * count);
         }
 
-        virtual void serializeObject(Object* object) = 0;
-        virtual void serializeBinary(void* data, size_t count) = 0;
-
-        virtual void serializeFundamental(refl::Fundamental* fundamental, void* data, size_t count = 1) override;
-        virtual void serializeStructure(refl::Structure* structure, void* data, size_t count = 1) override;
-
     private:
         Asset* mAsset;
-
         std::vector<std::string> mStringTable;
-
-        std::vector<ImportItem> mImportTable;
-
-        std::vector<ExportItem> mExportTable;
-
-        std::unordered_map<uint32_t, Object*> mTempObjects;
-
-        std::vector<ObjectBuffer> mObjectBufferTable;
-
+        std::unordered_map<std::string, uint32_t> mStringIndexMap;
+        std::vector<uint32_t> mImportTable;
         std::stack<uint32_t> mObjectBufferIndexStack;
+        std::vector<osg::ref_ptr<ObjectBuffer>> mObjectBuffers;
     };
 }
