@@ -10,6 +10,31 @@ namespace xxx::editor
 
         }
 
+        void drawTree(Entity* entity)
+        {
+            Context& context = Context::get();
+            const char* name = entity->getName().c_str();
+            size_t childrenCount = entity->getChildrenCount();
+            ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_DefaultOpen | ImGuiTreeNodeFlags_OpenOnArrow;
+            if (context.getActivedEntity() == entity)
+                flags |= ImGuiTreeNodeFlags_Selected;
+            if (!childrenCount)
+                flags |= ImGuiTreeNodeFlags_Leaf;
+            
+            if (ImGui::TreeNodeEx(name, flags))
+            {
+                if (ImGui::IsItemClicked())
+                {
+                    context.setActivedEntity(entity);
+                }
+                for (size_t i = 0; i < childrenCount; ++i)
+                {
+                    drawTree(entity->getChild(i));
+                }
+                ImGui::TreePop();
+            }
+        }
+
         virtual bool draw() override
         {
             if (!mVisibility)
@@ -17,7 +42,16 @@ namespace xxx::editor
 
             if (ImGui::Begin(mTitle.c_str(), &mVisibility))
             {
-
+                Scene* scene = Context::get().getScene();
+                if (scene)
+                {
+                    Entity* rootEntity = scene->getRootEntity();
+                    size_t childrenCount = rootEntity->getChildrenCount();
+                    for (size_t i = 0; i < childrenCount; ++i)
+                    {
+                        drawTree(rootEntity->getChild(i));
+                    }
+                }
             }
             ImGui::End();
 
