@@ -4,74 +4,81 @@
 
 namespace xxx::refl
 {
-    template <typename, typename = void>
-    struct is_comparable : std::false_type {};
-
-    template <typename T>
-    struct is_comparable<T, std::void_t<decltype(std::declval<const T>() == std::declval<const T>())>> : std::true_type {};
-
-    /*template <typename T>
-    struct is_comparable<T, std::void_t<decltype(std::declval<const T>() == std::declval<const T>()), std::enable_if_t<!is_template_instance<T>>>> : std::true_type {};
-
-    template <typename T1, typename T2>
-    struct is_comparable<std::map<T1, T2>, std::void_t<std::enable_if_t<is_comparable<T1>::value>, std::enable_if_t<is_comparable<T2>::value>>> : std::true_type {};
-
-    template <typename T1, typename T2>
-    struct is_comparable<std::pair<T1, T2>, std::void_t<std::enable_if_t<is_comparable<T1>::value>, std::enable_if_t<is_comparable<T2>::value>>> : std::true_type {};
-
-    template <typename T>
-    struct is_comparable<std::set<T>, std::enable_if_t<is_comparable<T>::value>> : std::true_type {};
-
-    template <typename T1, typename T2>
-    struct is_comparable<std::unordered_map<T1, T2>, std::void_t<std::enable_if_t<is_comparable<T1>::value>, std::enable_if_t<is_comparable<T2>::value>>> : std::true_type {};
-
-    template <typename T>
-    struct is_comparable<std::unordered_set<T>, std::enable_if_t<is_comparable<T>::value>> : std::true_type {};
-
-    template <typename T>
-    struct is_comparable<std::vector<T>, std::enable_if_t<is_comparable<T>::value>> : std::true_type {};*/
-
-    template <typename T>
-    static constexpr bool is_comparable_v = is_comparable<T>::value;
-
-    template <typename T>
-    struct array_element {
-        using type = T;
-    };
-
-    template <typename T, std::size_t Size>
-    struct array_element<T[Size]> {
-        using type = T;
-    };
-
-    template <typename T>
-    using array_element_t = typename array_element<T>::type;
-
 	class Property : public MetadataBase
 	{
 	public:
-        Property(std::string_view name) :
-            mName(name)
-        {}
-
         std::string_view getName() const
         {
             return mName;
         }
 
         virtual Type* getDeclaredType() const = 0;
-        virtual Type* getOwnerType() const = 0;
+        //virtual Type* getOwnerType() const = 0;
+        virtual void setValue(void* instance, const void* value) const = 0;
         virtual void getValue(const void* instance, void* value) const = 0;
         virtual void* getValuePtr(void* instance) const = 0;
         virtual const void* getValuePtr(const void* instance) const = 0;
-        virtual void setValue(void* instance, const void* value) const = 0;
         virtual bool compare(const void* instance1, const void* instance2) const = 0;
 
     protected:
+        Property(std::string_view name) :
+            mName(name) {}
+
+    private:
         std::string_view mName;
 	};
 
-    template <typename Owner, typename FakeDeclared, std::size_t Index = 0>
+    template <typename Declared, typename Owner>
+    class TProperty : public Property
+    {
+        using MemberPtr = Declared Owner::*;
+    public:
+        TProperty(std::string_view name, MemberPtr memberPtr) :
+            Property(name), mMemberPtr(memberPtr) {}
+
+        virtual Type* getDeclaredType() const override
+        {
+            return Reflection::getType<Declared>();
+        }
+
+        /*virtual Type* getOwnerType() const override
+        {
+            return Reflection::getType<Owner>();
+        }*/
+
+        virtual void setValue(void* instance, const void* value) const override
+        {
+            static_cast<Owner*>(instance)->*mMemberPtr = *static_cast<const Declared*>(value);
+        }
+
+        virtual void getValue(const void* instance, void* value) const override
+        {
+            *static_cast<Declared*>(value) = static_cast<const Owner*>(instance)->*mMemberPtr;
+        }
+
+        virtual void* getValuePtr(void* instance) const override
+        {
+            return &(static_cast<Owner*>(instance)->*mMemberPtr);
+        }
+
+        virtual const void* getValuePtr(const void* instance) const override
+        {
+            return &(static_cast<const Owner*>(instance)->*mMemberPtr);
+        }
+
+        virtual bool compare(const void* instance1, const void* instance2) const override
+        {
+            if constexpr (SpecialType<Declared> || !Comparable<Declared>)
+                return Reflection::getType<Declared>()->compare(getValuePtr(instance1), getValuePtr(instance2));
+            else
+                return *static_cast<const Declared*>(getValuePtr(instance1)) == *static_cast<const Declared*>(getValuePtr(instance2));
+        }
+
+    private:
+        MemberPtr mMemberPtr;
+    };
+
+    /*template <typename Owner, typename FakeDeclared, std::size_t Index = 0>
     class TProperty : public Property
     {
         using MemberObject = FakeDeclared Owner::*;
@@ -135,5 +142,5 @@ namespace xxx::refl
 
     protected:
         MemberObject mMemberObject;
-    };
+    };*/
 }
