@@ -5,121 +5,122 @@
 namespace xxx
 {
     using namespace refl;
-    void AssetLoader::serializeObject(Object* object)
-    {
-        Class* clazz = object->getClass();
 
-        uint32_t propertyCount;
-        serializeArithmetic(&propertyCount);
-
-        std::vector<Property*> properties;
-        Class* baseClass = clazz;
-        while (baseClass)
-        {
-            const std::vector<Property*>& props = baseClass->getProperties();
-            properties.insert(properties.end(), props.begin(), props.end());
-            baseClass = baseClass->getBaseClass();
-        }
-
-        for (size_t i = 0; i < propertyCount; ++i)
-        {
-            std::string propertyName;
-            uint32_t propertySize;
-            serializeStdString(&propertyName);
-            serializeArithmetic(&propertySize);
-
-            auto findResult = std::find_if(properties.begin(), properties.end(),
-                [propertyName](const Property* prop)->bool
-                {
-                return propertyName == prop->getName();
-                }
-            );
-            if (findResult != properties.end())
-            {
-                // found and serialize this property
-                Property* prop = *findResult;
-                properties.erase(findResult);
-
-                Type* declaredType = prop->getDeclaredType();
-                if (declaredType->getKind() == Type::Kind::Class)
-                {
-                    Object* propObject;
-                    serializeType(declaredType, &propObject);
-                    prop->setValue(object, &propObject);
-                }
-                else
-                {
-                    serializeType(declaredType, prop->getValuePtr(object));
-                }
-            }
-            else
-            {
-                // not found, skip this property
-                seek(tell() + propertySize);
-            }
-        }
-
-        object->postLoad();
-    }
-
-    void AssetLoader::serializeBinary(void* data, size_t count)
-    {
-        ObjectBuffer& objectBuffer = getCurrentObjectBuffer();
-        objectBuffer.readData(data, count);
-    }
+    AssetLoader::AssetLoader(Asset* asset) : AssetSerializer(asset) {}
 
     void AssetLoader::serializeEnumeration(Enumeration* enumeration, void* data, size_t count)
     {
         std::vector<std::string> valueNames(count);
-        serializeStdString(valueNames.data(), count);
+        serializeStdString(nullptr, valueNames.data(), count);
         for (size_t i = 0; i < count; ++i)
             enumeration->setValue(data, enumeration->getValueByName(valueNames[i]));
     }
 
-    void AssetLoader::serializeClass(Object** data, size_t count)
+    void AssetLoader::serializeClass(Class* clazz, void* data, size_t count)
     {
+        Object** objects = static_cast<Object**>(data);
         for (size_t i = 0; i < count; ++i)
         {
-            Object*& object = data[i];
-            int32_t index;
-            if (currentObjectBufferIsValid())
-                serializeArithmetic(&index);
+            Object*& object = objects[i];
+            int32_t objectIndex;
+            if (getCurrentObjectBuffer())
+                serializeArithmetic(&objectIndex);
             else
-                index = -1;
-            object = getObjectByIndex(index);
-        }
-    }
+                objectIndex = -1;
 
-    void AssetLoader::serializeStdString(std::string* data, size_t count)
-    {
-        std::vector<uint32_t> stringIndices(count);
-        serializeArithmetic(stringIndices.data(), count);
-        for (size_t i = 0; i < count; ++i)
-        {
-            data[i] = getStringTable().at(stringIndices[i]);
+            if (objectIndex == 0)
+            {
+                object = nullptr;
+            }
+            else if (objectIndex > 0) // Imported object
+            {
+                uint32_t importIndex = objectIndex - 1;
+                const std::string& path = getString(getImportItem(importIndex));
+                Asset* asset = AssetManager::get().getAsset(path);
+                if (asset)
+                {
+                    object = asset->getRootObjectSafety();
+                }
+                else
+                {
+                    //LOG_ERROR("Cannot find asset: {}", path);
+                    object = nullptr;
+                }
+            }
+            else // Exported object
+            {
+                uint32_t exportIndex = -objectIndex - 1;
+                auto objectFindResult = mObjectsTemp.find(exportIndex);
+                if (objectFindResult == mObjectsTemp.end())
+                {
+                    pushObjectBuffer(exportIndex);
+                    std::string className;
+                    serializeStdString(nullptr, &className);
+                    Class* clazz = Reflection::getClass(className);
+                    object = static_cast<Object*>(clazz->newInstance());
+                    mObjectsTemp.emplace(exportIndex, object);
+                    serializeObject(object);
+                    popObjectBuffer();
+                }
+                else
+                {
+                    object = objectFindResult->second;
+                }
+            }
         }
     }
 
     void AssetLoader::serializeStdArray(StdArray* stdArray, void* data, size_t count)
     {
-        const size_t stdArraySize = stdArray->getSize();
+        Type* elementType = stdArray->getElementType();
+        size_t elementCount = stdArray->getElementCount();
         for (size_t i = 0; i < count; ++i)
         {
-            void* stdArrayData = static_cast<uint8_t*>(data) + stdArraySize * i;
-            Type* elementType = stdArray->getElementType();
-            size_t elementCount = stdArray->getElementCount();
+            void* stdArrayData = static_cast<uint8_t*>(data) + stdArray->getSize() * i;
             if (elementType->getKind() == Type::Kind::Class)
             {
-                Object** objects = new Object*[elementCount];
-                serializeType(stdArray->getElementType(), objects, elementCount);
+                Object** objects = new Object * [elementCount];
+                serializeClass(dynamic_cast<Class*>(elementType), objects, elementCount);
                 for (size_t j = 0; j < elementCount; ++j)
-                    stdArray->setElementValue(stdArrayData, j, &objects[j]);
+                    stdArray->setElement(stdArrayData, j, &objects[j]);
                 delete[] objects;
             }
             else
             {
-                serializeType(elementType, stdArray->getElementPtrByIndex(stdArrayData, 0), elementCount);
+                serializeType(elementType, stdArray->getElementPtr(stdArrayData, 0), elementCount);
             }
+        }
+    }
+
+    void AssetLoader::serializeStdList(StdList* stdList, void* data, size_t count)
+    {
+        Type* elementType = stdList->getElementType();
+        const bool elementTypeIsClass = elementType->getKind() == Type::Kind::Class;
+        const size_t elementSize = elementTypeIsClass ? sizeof(Object*) : elementType->getSize();
+        for (size_t i = 0; i < count; ++i)
+        {
+            void* stdListData = static_cast<uint8_t*>(data) + stdList->getSize() * i;
+            size_t elementCount;
+            serializeArithmetic(&elementCount);
+
+            if (elementCount == 0)
+                continue;
+
+            void* elements = elementTypeIsClass ?
+                new Object * [elementCount] :
+                elementType->newInstances(elementCount);
+            serializeType(elementType, elements, elementCount);
+
+            for (size_t j = 0; j < elementCount; ++j)
+            {
+                void* element = static_cast<uint8_t*>(elements) + j * elementSize;
+                stdList->addElement(stdListData, element);
+            }
+
+            if (elementTypeIsClass)
+                delete[] static_cast<Object**>(elements);
+            else
+                elementType->deleteInstances(elements);
         }
     }
 
@@ -127,14 +128,19 @@ namespace xxx
     {
         Type* keyType = stdMap->getKeyType();
         Type* valueType = stdMap->getValueType();
+        const bool keyTypeIsClass = keyType->getKind() == Type::Kind::Class;
+        const bool valueTypeIsClass = valueType->getKind() == Type::Kind::Class;
+        const size_t keySize = keyTypeIsClass ? sizeof(Object*) : keyType->getSize();
+        const size_t valueSize = valueTypeIsClass ? sizeof(Object*) : valueType->getSize();
         for (size_t i = 0; i < count; ++i)
         {
             void* stdMapData = static_cast<uint8_t*>(data) + stdMap->getSize() * i;
             size_t keyValuePairCount;
             serializeArithmetic(&keyValuePairCount);
 
-            const bool keyTypeIsClass = keyType->getKind() == Type::Kind::Class;
-            const bool valueTypeIsClass = valueType->getKind() == Type::Kind::Class;
+            if (keyValuePairCount == 0)
+                continue;
+
             void* keys = keyTypeIsClass ?
                 new Object * [keyValuePairCount] :
                 keyType->newInstances(keyValuePairCount);
@@ -144,8 +150,6 @@ namespace xxx
             serializeType(keyType, keys, keyValuePairCount);
             serializeType(valueType, values, keyValuePairCount);
 
-            const size_t keySize = keyTypeIsClass ? sizeof(Object*) : keyType->getSize();
-            const size_t valueSize = valueTypeIsClass ? sizeof(Object*) : valueType->getSize();
             for (size_t j = 0; j < keyValuePairCount; ++j)
             {
                 void* key = static_cast<uint8_t*>(keys) + j * keySize;
@@ -166,14 +170,16 @@ namespace xxx
 
     void AssetLoader::serializeStdPair(StdPair* stdPair, void* data, size_t count)
     {
+        Type* firstType = stdPair->getFirstType();
+        Type* secondType = stdPair->getSecondType();
         for (size_t i = 0; i < count; ++i)
         {
             void* stdPairData = static_cast<uint8_t*>(data) + stdPair->getSize() * i;
-            Type* firstType = stdPair->getFirstType();
+
             if (firstType->getKind() == Type::Kind::Class)
             {
                 Object* object;
-                serializeType(firstType, &object);
+                serializeClass(dynamic_cast<Class*>(firstType), &object);
                 stdPair->setFirstValue(stdPairData, &object);
             }
             else
@@ -181,11 +187,10 @@ namespace xxx
                 serializeType(firstType, stdPair->getFirstPtr(stdPairData));
             }
 
-            Type* secondType = stdPair->getSecondType();
             if (secondType->getKind() == Type::Kind::Class)
             {
                 Object* object;
-                serializeType(secondType, &object);
+                serializeClass(dynamic_cast<Class*>(secondType), &object);
                 stdPair->setSecondValue(stdPairData, &object);
             }
             else
@@ -198,19 +203,22 @@ namespace xxx
     void AssetLoader::serializeStdSet(StdSet* stdSet, void* data, size_t count)
     {
         Type* elementType = stdSet->getElementType();
+        const bool elementTypeIsClass = elementType->getKind() == Type::Kind::Class;
+        const size_t elementSize = elementTypeIsClass ? sizeof(Object*) : elementType->getSize();
         for (size_t i = 0; i < count; ++i)
         {
             void* stdSetData = static_cast<uint8_t*>(data) + stdSet->getSize() * i;
             size_t elementCount;
             serializeArithmetic(&elementCount);
 
-            const bool elementTypeIsClass = elementType->getKind() == Type::Kind::Class;
+            if (elementCount == 0)
+                continue;
+
             void* elements = elementTypeIsClass ?
                 new Object * [elementCount] :
                 elementType->newInstances(elementCount);
             serializeType(elementType, elements, elementCount);
 
-            const size_t elementSize = elementTypeIsClass ? sizeof(Object*) : elementType->getSize();
             for (size_t j = 0; j < elementCount; ++j)
             {
                 void* element = static_cast<uint8_t*>(elements) + j * elementSize;
@@ -224,28 +232,12 @@ namespace xxx
         }
     }
 
-    void AssetLoader::serializeStdTuple(StdTuple* stdTuple, void* data, size_t count)
+    void AssetLoader::serializeStdString(StdString* stdString, void* data, size_t count)
     {
+        std::vector<uint32_t> strIndices(count);
+        serializeArithmetic(strIndices.data(), count);
         for (size_t i = 0; i < count; ++i)
-        {
-            void* stdTupleData = static_cast<uint8_t*>(data) + stdTuple->getSize() * i;
-            std::vector<Type*> tupleTypes = stdTuple->getTypes();
-            std::vector<void*> tupleElementPtrs = stdTuple->getElementPtrs(stdTupleData);
-            size_t tupleElementCount = stdTuple->getElementCount();
-            for (size_t j = 0; j < tupleElementCount; ++j)
-            {
-                if (tupleTypes[j]->getKind() == Type::Kind::Class)
-                {
-                    Object* object;
-                    serializeType(tupleTypes[j], &object);
-                    stdTuple->setElementValue(stdTupleData, j, &object);
-                }
-                else
-                {
-                    serializeType(tupleTypes[j], tupleElementPtrs[j]);
-                }
-            }
-        }
+            static_cast<std::string*>(data)[i] = getString(strIndices[i]);
     }
 
     void AssetLoader::serializeStdUnorderedMap(refl::StdUnorderedMap* stdUnorderedMap, void* data, size_t count)
@@ -320,23 +312,24 @@ namespace xxx
 
     void AssetLoader::serializeStdVariant(StdVariant* stdVariant, void* data, size_t count)
     {
+        std::vector<Type*> types = stdVariant->getTypes();
         for (size_t i = 0; i < count; ++i)
         {
             void* stdVariantData = static_cast<uint8_t*>(data) + stdVariant->getSize() * i;
             uint32_t typeIndex;
             serializeArithmetic(&typeIndex);
-            Type* variantType = stdVariant->getTypes().at(typeIndex);
+            Type* variantType = types.at(typeIndex);
             if (variantType->getKind() == Type::Kind::Class)
             {
                 Object* object;
-                serializeType(variantType, &object);
-                stdVariant->setValueByTypeIndex(stdVariantData, typeIndex, &object);
+                serializeClass(dynamic_cast<Class*>(variantType), &object);
+                stdVariant->setValue(stdVariantData, typeIndex, &object);
             }
             else
             {
                 void* valuePtr = variantType->newInstance();
                 serializeType(variantType, valuePtr);
-                stdVariant->setValueByTypeIndex(stdVariantData, typeIndex, valuePtr);
+                stdVariant->setValue(stdVariantData, typeIndex, valuePtr);
                 variantType->deleteInstance(valuePtr);
             }
         }
@@ -350,22 +343,85 @@ namespace xxx
             void* stdVectorData = static_cast<uint8_t*>(data) + stdVector->getSize() * i;
             size_t elementCount;
             serializeArithmetic(&elementCount);
-            if (elementCount > 0)
+
+            if (elementCount == 0)
+                continue;
+
+            stdVector->resize(stdVectorData, elementCount);
+            if (elementType->getKind() == Type::Kind::Class)
             {
-                stdVector->resize(stdVectorData, elementCount);
-                if (elementType->getKind() == Type::Kind::Class)
+                Object** objects = new Object * [elementCount];
+                serializeClass(dynamic_cast<Class*>(elementType), objects, elementCount);
+                for (size_t j = 0; j < elementCount; ++j)
+                    stdVector->setElement(stdVectorData, j, &objects[j]);
+                delete[] objects;
+            }
+            else
+            {
+                serializeType(elementType, stdVector->getElementPtr(stdVectorData, 0), elementCount);
+            }
+        }
+    }
+
+    void AssetLoader::serializeBinary(void* data, size_t size)
+    {
+        getCurrentObjectBuffer()->read(data, size);
+    }
+
+    void AssetLoader::serializeObject(Object* object)
+    {
+        ObjectBuffer* objectBuffer = getCurrentObjectBuffer();
+
+        Class* clazz = object->getClass();
+
+        uint32_t propertyCount;
+        serializeArithmetic(&propertyCount);
+
+        std::vector<Property*> properties;
+        Class* baseClass = clazz;
+        while (baseClass)
+        {
+            const std::vector<Property*>& props = baseClass->getProperties();
+            properties.insert(properties.end(), props.begin(), props.end());
+            baseClass = baseClass->getBaseClass();
+        }
+
+        for (size_t i = 0; i < propertyCount; ++i)
+        {
+            std::string propertyName;
+            uint32_t propertySize;
+            serializeStdString(nullptr, &propertyName);
+            serializeArithmetic(&propertySize);
+
+            auto findResult = std::find_if(properties.begin(), properties.end(),
+                [propertyName](const Property* prop)->bool { return propertyName == prop->getName(); }
+            );
+
+            if (findResult != properties.end())
+            {
+                // found and serialize this property
+                Property* prop = *findResult;
+                properties.erase(findResult);
+
+                Type* declaredType = prop->getDeclaredType();
+                if (declaredType->getKind() == Type::Kind::Class)
                 {
-                    Object** objects = new Object * [elementCount];
-                    serializeType(elementType, objects, elementCount);
-                    for (size_t j = 0; j < elementCount; ++j)
-                        stdVector->setElementValue(stdVectorData, j, &objects[j]);
-                    delete[] objects;
+                    Object* propObject;
+                    serializeType(declaredType, &propObject);
+                    prop->setValue(object, &propObject);
                 }
                 else
                 {
-                    serializeType(elementType, stdVector->getElementPtrByIndex(stdVectorData, 0), elementCount);
+                    serializeType(declaredType, prop->getValuePtr(object));
                 }
             }
+            else
+            {
+                // not found, skip this property
+                objectBuffer->seek(objectBuffer->tell() + propertySize);
+            }
         }
+
+        object->postLoad();
     }
 }
