@@ -5,95 +5,89 @@ namespace xxx
 {
     using namespace refl;
 
-    void AssetSaver::serializeObject(Object* object)
-    {
-        object->preSave();
-
-        Class* clazz = object->getClass();
-        const Object* defaultObject = clazz->getDefaultObject();
-        std::vector<Property*> properties;
-        Class* baseClass = clazz;
-        while (baseClass)
-        {
-            for (Property* prop : baseClass->getProperties())
-            {
-                if (!prop->compare(defaultObject, object))
-                    properties.emplace_back(prop);
-            }
-            baseClass = baseClass->getBaseClass();
-        }
-        uint32_t propertyCount = properties.size();
-        serializeArithmetic(&propertyCount);
-
-        for (Property* prop : properties)
-        {
-            std::string propertyName(prop->getName());
-            uint32_t propertySize = 0;
-
-            serializeStdString(&propertyName);
-            uint32_t propertySizePos = tell();
-            serializeArithmetic(&propertySize);
-
-            uint32_t propertyBeginPos = tell();
-            void* valuePtr = prop->getValuePtr(object);
-            serializeType(prop->getDeclaredType(), valuePtr);
-            uint32_t propertyEndPos = tell();
-
-            propertySize = propertyEndPos - propertyBeginPos;
-            seek(propertySizePos);
-            serializeArithmetic(&propertySize);
-            seek(propertyEndPos);
-        }
-
-        object->postSave();
-    }
-
-    void AssetSaver::serializeBinary(void* data, size_t count)
-    {
-        ObjectBuffer& objectBuffer = getCurrentObjectBuffer();
-        objectBuffer.writeData(data, count);
-    }
+    AssetSaver::AssetSaver(Asset* asset) : AssetSerializer(asset) {}
 
     void AssetSaver::serializeEnumeration(Enumeration* enumeration, void* data, size_t count)
     {
         std::vector<std::string> valueNames(count);
         for (size_t i = 0; i < count; ++i)
             valueNames[i] = enumeration->getNameByValue(enumeration->getValue(data));
-        serializeStdString(valueNames.data(), count);
+        serializeStdString(nullptr, valueNames.data(), count);
     }
 
-    void AssetSaver::serializeClass(Object** data, size_t count)
+    void AssetSaver::serializeClass(Class* clazz, void* data, size_t count)
     {
+        Object** objects = static_cast<Object**>(data);
         for (size_t i = 0; i < count; ++i)
         {
-            Object* object = data[i];
-
-            int32_t index = getIndexOfObject(object);
-
-            if (currentObjectBufferIsValid())
+            Object* object = objects[i];
+            int32_t objectIndex;
+            if (object == nullptr)
+                objectIndex = 0;
+            else
             {
-                serializeArithmetic(&index);
-            }
-        }
-    }
+                Asset* asset = object->getAsset();
+                if (asset != getAsset() && asset != nullptr)
+                {
+                    if (object->getOwner() != nullptr)
+                    {
+                        //LOG_ERROR("Cannot reference object that is not asset root object");
+                        objectIndex = 0;
+                    }
+                    else
+                    {
+                        uint32_t importIndex = addImportItem(addString(asset->getPath()));
+                        objectIndex = importIndex + 1;
+                    }
+                }
+                else
+                {
+                    uint32_t exportIndex;
+                    auto exportIndexFindResult = mExportIndexTemp.find(object);
+                    if (exportIndexFindResult == mExportIndexTemp.end())
+                    {
+                        exportIndex = createObjectBuffer();
+                        mExportIndexTemp.emplace(object, exportIndex);
 
-    void AssetSaver::serializeStdString(std::string* data, size_t count)
-    {
-        std::vector<uint32_t> stringIndices(count);
-        for (size_t i = 0; i < count; ++i)
-        {
-            stringIndices[i] = addString(data[i]);
+                        pushObjectBuffer(exportIndex);
+                        std::string className(object->getClass()->getName());
+                        serializeStdString(nullptr, &className);
+                        serializeObject(object);
+                        popObjectBuffer();
+                    }
+                    else
+                    {
+                        exportIndex = exportIndexFindResult->second;
+                    }
+                    objectIndex = -int32_t(exportIndex + 1);
+                }
+            }
+            if (getCurrentObjectBuffer())
+                serializeArithmetic(&objectIndex);
         }
-        serializeArithmetic(stringIndices.data(), count);
     }
 
     void AssetSaver::serializeStdArray(StdArray* stdArray, void* data, size_t count)
     {
-        const size_t stdArraySize = stdArray->getSize();
+        Type* elementType = stdArray->getElementType();
         for (size_t i = 0; i < count; ++i)
         {
-            void* stdArrayData = static_cast<uint8_t*>(data) + stdArraySize * i;
-            serializeType(stdArray->getElementType(), stdArray->getElementPtrByIndex(stdArrayData, 0), stdArray->getElementCount());
+            void* stdArrayData = static_cast<uint8_t*>(data) + stdArray->getSize() * i;
+            serializeType(elementType, stdArray->getElementPtr(stdArrayData, 0), stdArray->getElementCount());
+        }
+    }
+
+    void AssetSaver::serializeStdList(StdList* stdList, void* data, size_t count)
+    {
+        Type* elementType = stdList->getElementType();
+        for (size_t i = 0; i < count; ++i)
+        {
+            void* stdListData = static_cast<uint8_t*>(data) + stdList->getSize() * i;
+            std::vector<void*> elementPtrs = stdList->getElementPtrs(stdListData);
+            size_t elementCount = elementPtrs.size();
+            serializeArithmetic(&elementCount);
+            for (size_t j = 0; j < elementCount; ++j)
+                serializeType(elementType, static_cast<void*>(elementPtrs[j]));
         }
     }
 
@@ -104,13 +98,14 @@ namespace xxx
         for (size_t i = 0; i < count; ++i)
         {
             void* stdMapData = static_cast<uint8_t*>(data) + stdMap->getSize() * i;
-            std::vector<std::pair<const void*, void*>> keyValuePtrs = stdMap->getKeyValuePtrs(stdMapData);
-            size_t keyValuePairCount = keyValuePtrs.size();
+            size_t keyValuePairCount = stdMap->getKeyValuePairCount(stdMapData);
             serializeArithmetic(&keyValuePairCount);
+            std::vector<const void*> keys = stdMap->getKeyPtrs(stdMapData);
             for (size_t j = 0; j < keyValuePairCount; ++j)
-                serializeType(keyType, const_cast<void*>(keyValuePtrs[j].first));
+                serializeType(keyType, const_cast<void*>(keys[j]));
+            std::vector<void*> values = stdMap->getValuePtrs(stdMapData);
             for (size_t j = 0; j < keyValuePairCount; ++j)
-                serializeType(valueType, keyValuePtrs[j].second);
+                serializeType(valueType, values[j]);
         }
     }
 
@@ -138,17 +133,12 @@ namespace xxx
         }
     }
 
-    void AssetSaver::serializeStdTuple(StdTuple* stdTuple, void* data, size_t count)
+    void AssetSaver::serializeStdString(StdString* stdString, void* data, size_t count)
     {
+        std::vector<uint32_t> stringIndices(count);
         for (size_t i = 0; i < count; ++i)
-        {
-            void* stdTupleData = static_cast<uint8_t*>(data) + stdTuple->getSize() * i;
-            std::vector<Type*> tupleTypes = stdTuple->getTypes();
-            std::vector<void*> tupleElementPtrs = stdTuple->getElementPtrs(stdTupleData);
-            size_t tupleElementCount = stdTuple->getElementCount();
-            for (size_t i = 0; i < tupleElementCount; ++i)
-                serializeType(tupleTypes[i], tupleElementPtrs[i]);
-        }
+            stringIndices[i] = addString(static_cast<std::string*>(data)[i]);
+        serializeArithmetic(stringIndices.data(), count);
     }
 
     void AssetSaver::serializeStdUnorderedMap(refl::StdUnorderedMap* stdUnorderedMap, void* data, size_t count)
@@ -158,13 +148,14 @@ namespace xxx
         for (size_t i = 0; i < count; ++i)
         {
             void* stdUnorderedMapData = static_cast<uint8_t*>(data) + stdUnorderedMap->getSize() * i;
-            std::vector<std::pair<const void*, void*>> keyValuePtrs = stdUnorderedMap->getKeyValuePtrs(stdUnorderedMapData);
-            size_t keyValuePairCount = keyValuePtrs.size();
+            size_t keyValuePairCount = stdUnorderedMap->getKeyValuePairCount(stdUnorderedMapData);
             serializeArithmetic(&keyValuePairCount);
+            std::vector<const void*> keys = stdUnorderedMap->getKeyPtrs(stdUnorderedMapData);
             for (size_t j = 0; j < keyValuePairCount; ++j)
-                serializeType(keyType, const_cast<void*>(keyValuePtrs[j].first));
+                serializeType(keyType, const_cast<void*>(keys[j]));
+            std::vector<void*> values = stdUnorderedMap->getValuePtrs(stdUnorderedMapData);
             for (size_t j = 0; j < keyValuePairCount; ++j)
-                serializeType(valueType, keyValuePtrs[j].second);
+                serializeType(valueType, values[j]);
         }
     }
 
@@ -204,7 +195,57 @@ namespace xxx
             size_t elementCount = stdVector->getElementCount(stdVectorData);
             serializeArithmetic(&elementCount);
             if (elementCount > 0)
-                serializeType(elementType, stdVector->getElementPtrByIndex(stdVectorData, 0), elementCount);
+                serializeType(elementType, stdVector->getElementPtr(stdVectorData, 0), elementCount);
         }
+    }
+
+    void AssetSaver::serializeBinary(void* data, size_t size)
+    {
+        getCurrentObjectBuffer()->write(data, size);
+    }
+
+    void AssetSaver::serializeObject(Object* object)
+    {
+        object->preSave();
+
+        ObjectBuffer* objectBuffer = getCurrentObjectBuffer();
+
+        Class* clazz = object->getClass();
+        const Object* defaultObject = clazz->getDefaultObject();
+        std::vector<Property*> properties;
+        Class* baseClass = clazz;
+        while (baseClass)
+        {
+            for (Property* prop : baseClass->getProperties())
+            {
+                if (!prop->compare(defaultObject, object))
+                    properties.emplace_back(prop);
+            }
+            baseClass = baseClass->getBaseClass();
+        }
+        uint32_t propertyCount = properties.size();
+        serializeArithmetic(&propertyCount);
+
+        for (Property* prop : properties)
+        {
+            std::string propertyName(prop->getName());
+            uint32_t propertySize = 0;
+
+            serializeStdString(nullptr, &propertyName);
+            uint32_t propertySizePos = objectBuffer->tell();
+            serializeArithmetic(&propertySize);
+
+            uint32_t propertyBeginPos = objectBuffer->tell();
+            void* valuePtr = prop->getValuePtr(object);
+            serializeType(prop->getDeclaredType(), valuePtr);
+            uint32_t propertyEndPos = objectBuffer->tell();
+
+            propertySize = propertyEndPos - propertyBeginPos;
+            objectBuffer->seek(propertySizePos);
+            serializeArithmetic(&propertySize);
+            objectBuffer->seek(propertyEndPos);
+        }
+
+        object->postSave();
     }
 }
